requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Jinja2==3.1.4
Werkzeug==3.0.3
WTForms==3.1.2
WTForms-SQLAlchemy==0.4.1
email-validator==2.1.1
python-dotenv==1.0.1